--- packages/ui/src/index.ts
@@ -1,2 +1,0 @@
-export { default as i18n } from './services/i18n';
-export * from './stores/notifications-store';
