+++ packages/ui/src/legacy/components/NotificationHub/NotificationHub.tsx
@@ -2,7 +2,7 @@ import { XMarkIcon } from '@heroicons/react/24/solid';
 import { AnimatePresence, motion } from 'framer-motion';
 import { useTranslation } from 'react-i18next';
 
-import { useNotificationsStore } from '@/stores/notifications-store';
+import { useNotificationsStore } from '@/hooks/useNotificationsStore';
 
 import { Card } from '../Card/Card';
 import { NotificationIcon } from './NotificationIcon';
